requirements.txt
# Core dependencies for stock prediction application
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
vaderSentiment==3.3.2
yfinance==0.2.28
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
scikit-learn==1.3.2
nltk==3.8.1
newsapi-python==0.2.6
tweepy==4.14.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
matplotlib==3.8.2
GoogleNews==1.6.12
apscheduler==3.10.4
ta-lib==0.4.28

# Additional missing dependencies
alpha-vantage==2.3.1
prophet==1.1.4
arch==6.2.0
ruptures==1.1.7
lightgbm==4.1.0
xgboost==2.0.3
shap==0.44.0
feedparser==6.0.10
statsmodels==0.14.0
networkx==3.2.1